ref(testing): move 'testing' feature to base
