+++ react/features/base/testing/components/TestHint.android.js
