+++ react/features/base/testing/components/TestHint.ios.js
